removed auth model
